@@ -21,10 +21,10 @@
     "js-yaml": "^3.14.0",
     "prop-types": "^15.7.2",
     "rc-slider": "^9.6.1",
-    "react": "^17.0.1",
+    "react": "^18.3.1",
     "react-bootstrap": "^1.4.0",
     "react-color": "^2.19.3",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^0.2.4"
   },
